+++ source/modules/tgv_cuda/tgv_common.cu
@@ -208,7 +208,7 @@ __global__ void tgv_kernel_part2(
             norm3df(p_xx[index], p_yy[index], p_zz[index]) :
             sqrtf(p_xx[index] * p_xx[index] + p_yy[index] * p_yy[index]);
 
-    normalization = fmax(1, normalization/alpha1);
+    normalization = fmaxf(1, normalization/alpha1);
 
     p_xx[index] /= normalization;
     p_yy[index] /= normalization;
